refactor: AI 프롬프트 생성 책임 분리
- 강의 자료 AI 프롬프트 Builder 분리

- 학습 자료 OCR 프롬프트 Builder 분리

- 공통 AI 프롬프트 record 추가

- 기존 프로세서의 프롬프트 생성 로직 제거

- 테스트 생성자 의존성 반영
