+++ src/main/java/com/f1/quiket/domain/lecture/service/LectureMaterialAiProcessor.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.f1.quiket.domain.lecture.dto.LectureMaterialAiProcessRequest;
 import com.f1.quiket.domain.lecture.dto.LectureMaterialAiProcessResult;
 import com.f1.quiket.domain.lecture.dto.LecturePartDraft;
-import com.f1.quiket.domain.lecture.dto.LecturePartSplitPlan;
+import com.f1.quiket.domain.material.dto.StudyMaterialAiPrompt;
 import com.f1.quiket.domain.material.dto.StudyMaterialFile;
 import com.f1.quiket.domain.material.dto.StudyMaterialPdfTextExtractionResult;
 import com.f1.quiket.domain.material.dto.StudyMaterialUploadType;
@@ -14,9 +14,7 @@ import com.f1.quiket.domain.material.port.StudyMaterialPdfTextExtractor;
 import com.f1.quiket.global.error.CustomException;
 import com.f1.quiket.global.response.ErrorCode;
 import java.util.ArrayList;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
@@ -32,14 +30,9 @@ import org.springframework.util.StringUtils;
 @RequiredArgsConstructor
 public class LectureMaterialAiProcessor {
 
-    private static final String SYSTEM_MESSAGE = """
-            너는 Quiket 강의 자료 파트 분류 엔진이다.
-            반드시 JSON만 반환한다.
-            코드블록, 마크다운, 설명 문장은 절대 포함하지 않는다.
-            """;
-
     private final StudyMaterialAiGateway studyMaterialAiGateway;
     private final StudyMaterialPdfTextExtractor studyMaterialPdfTextExtractor;
+    private final LectureMaterialAiPromptBuilder promptBuilder;
     private final ObjectMapper objectMapper = new ObjectMapper();
 
     /**
@@ -60,10 +53,11 @@ public class LectureMaterialAiProcessor {
      * 이미지 OCR 및 파트 분류 처리
      */
     private LectureMaterialAiProcessResult processImage(LectureMaterialAiProcessRequest request) {
+        StudyMaterialAiPrompt prompt = promptBuilder.buildGeminiPrompt(request, null);
         // Gemini 기반 이미지 OCR 및 파트 분류
         String aiResponse = studyMaterialAiGateway.generateFromImages(
-                SYSTEM_MESSAGE,
-                buildGeminiPrompt(request, null),
+                prompt.systemMessage(),
+                prompt.userMessage(),
                 request.getFiles()
         );
         List<LecturePartDraft> parts = parseParts(aiResponse);
@@ -84,9 +78,10 @@ public class LectureMaterialAiProcessor {
         // 텍스트 레이어 PDF는 추출 텍스트 기반 Groq 분류
         if (extractionResult.isHasTextLayer()) {
             String text = extractionResult.getExtractedText();
+            StudyMaterialAiPrompt prompt = promptBuilder.buildGroqPrompt(request, text);
             String aiResponse = studyMaterialAiGateway.generateFromText(
-                    SYSTEM_MESSAGE,
-                    buildGroqPrompt(request, text)
+                    prompt.systemMessage(),
+                    prompt.userMessage()
             );
             return LectureMaterialAiProcessResult.builder()
                     .provider("groq")
@@ -95,10 +90,11 @@ public class LectureMaterialAiProcessor {
                     .build();
         }
 
+        StudyMaterialAiPrompt prompt = promptBuilder.buildGeminiPrompt(request, null);
         // 스캔 PDF는 Gemini OCR 및 파트 분류
         String aiResponse = studyMaterialAiGateway.generateFromPdf(
-                SYSTEM_MESSAGE,
-                buildGeminiPrompt(request, null),
+                prompt.systemMessage(),
+                prompt.userMessage(),
                 pdfFile
         );
         return LectureMaterialAiProcessResult.builder()
@@ -112,10 +108,11 @@ public class LectureMaterialAiProcessor {
      * 직접 입력 텍스트 기반 파트 분류 처리
      */
     private LectureMaterialAiProcessResult processText(LectureMaterialAiProcessRequest request) {
+        StudyMaterialAiPrompt prompt = promptBuilder.buildGroqPrompt(request, request.getText());
         // 입력 텍스트 기반 Groq 파트 분류
         String aiResponse = studyMaterialAiGateway.generateFromText(
-                SYSTEM_MESSAGE,
-                buildGroqPrompt(request, request.getText())
+                prompt.systemMessage(),
+                prompt.userMessage()
         );
         return LectureMaterialAiProcessResult.builder()
                 .provider("groq")
@@ -138,88 +135,6 @@ public class LectureMaterialAiProcessor {
                 && (request.getFiles() == null || request.getFiles().isEmpty())) {
             throw new CustomException(ErrorCode.INVALID_INPUT_VALUE, "파일 업로드는 최소 1개 파일이 필요합니다.");
         }
-    }
-
-    /**
-     * Groq 텍스트 분류 프롬프트 생성
-     */
-    private String buildGroqPrompt(LectureMaterialAiProcessRequest request, String sourceText) {
-        return """
-                [입력]
-                - 챕터명: %s
-                - 파트 분류 방식: %s
-
-                [분류 계획]
-                %s
-
-                [텍스트]
-                %s
-
-                [반환 규칙]
-                - 반환 형식: {"parts":[{"partNumber":1,"name":"...","content":"..."}]}
-                - partNumber는 1부터 시작하는 오름차순 정수
-                - name은 1자 이상 30자 이하
-                - content는 핵심 내용을 유지한 한국어 본문
-                - partSplitMethod가 manual이면 분류 계획의 partNumber를 그대로 사용
-                - partSplitMethod가 manual이고 intendedName이 있으면 반드시 그 이름 사용
-                - parts는 최소 1개 이상 생성
-                """.formatted(
-                valueOrDefault(request.getChapterName(), "미지정"),
-                request.getPartSplitMethod().getValue(),
-                partSplitPlanContext(request.getPartSplitPlans()),
-                valueOrDefault(sourceText, "")
-        );
-    }
-
-    /**
-     * Gemini OCR 및 분류 프롬프트 생성
-     */
-    private String buildGeminiPrompt(LectureMaterialAiProcessRequest request, String sourceText) {
-        return """
-                [작업]
-                업로드 파일에서 텍스트를 추출(OCR 포함)한 뒤 파트를 분류한다.
-
-                [입력]
-                - 챕터명: %s
-                - 파트 분류 방식: %s
-
-                [분류 계획]
-                %s
-
-                [추가 텍스트]
-                %s
-
-                [반환 규칙]
-                - 반환 형식: {"parts":[{"partNumber":1,"name":"...","content":"..."}]}
-                - partNumber는 1부터 시작하는 오름차순 정수
-                - name은 1자 이상 30자 이하
-                - content는 핵심 내용을 유지한 한국어 본문
-                - partSplitMethod가 manual이면 분류 계획의 partNumber를 그대로 사용
-                - partSplitMethod가 manual이고 intendedName이 있으면 반드시 그 이름 사용
-                - parts는 최소 1개 이상 생성
-                """.formatted(
-                valueOrDefault(request.getChapterName(), "미지정"),
-                request.getPartSplitMethod().getValue(),
-                partSplitPlanContext(request.getPartSplitPlans()),
-                valueOrDefault(sourceText, "")
-        );
-    }
-
-    /**
-     * 수동 파트 분류 계획 프롬프트 컨텍스트 생성
-     */
-    private String partSplitPlanContext(List<LecturePartSplitPlan> plans) {
-        if (plans == null || plans.isEmpty()) {
-            return "- 계획 없음";
-        }
-        // 파트 번호 순서 기반 계획 정렬
-        return plans.stream()
-                .sorted(Comparator.comparing(LecturePartSplitPlan::getPartNumber))
-                .map(plan -> "- partNumber: %d, intendedName: %s".formatted(
-                        plan.getPartNumber(),
-                        valueOrDefault(plan.getIntendedName(), "미지정")
-                ))
-                .collect(Collectors.joining("\n"));
     }
 
     /**
