+++ src/main/java/com/f1/quiket/domain/material/dto/StudyMaterialAiPrompt.java
@@ -1,0 +1,7 @@
+package com.f1.quiket.domain.material.dto;
+
+public record StudyMaterialAiPrompt(
+        String systemMessage,
+        String userMessage
+) {
+}
