+++ src/test/java/com/f1/quiket/domain/lecture/service/LectureMaterialAiProcessorTest.java
@@ -31,7 +31,8 @@ class LectureMaterialAiProcessorTest {
         studyMaterialPdfTextExtractor = Mockito.mock(StudyMaterialPdfTextExtractor.class);
         lectureMaterialAiProcessor = new LectureMaterialAiProcessor(
                 studyMaterialAiGateway,
-                studyMaterialPdfTextExtractor
+                studyMaterialPdfTextExtractor,
+                new LectureMaterialAiPromptBuilder()
         );
     }
